% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{computePairIBD12}
\alias{computePairIBD12}
\title{Computes pairwise IBD1/2 for a specific pair of individuals}
\usage{
computePairIBD12(i, j)
}
\arguments{
\item{i}{Index of first individual}

\item{j}{Index of second individual}
}
\value{
Mean IBD1 and IBD2 as computed from shared haplotypes
}
\description{
Computes pairwise IBD1/2 for a specific pair of individuals
}

