% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{computePairIBD2}
\alias{computePairIBD2}
\title{Computes pairwise IBD2 for a specific pair of individuals}
\usage{
computePairIBD2(i, j)
}
\arguments{
\item{i}{Index of first individual}

\item{j}{Index of second individual}
}
\value{
Mean IBD2 as computed from shared haplotypes
}
\description{
Computes pairwise IBD2 for a specific pair of individuals
}

