% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readvcf.R
\name{createVCF}
\alias{createVCF}
\title{Creates a regional vcf file using bcftools to extract a region from 1000 genomes vcf files}
\usage{
createVCF()
}
\value{
none
}
\description{
Creates a regional vcf file using bcftools to extract a region from 1000 genomes vcf files
}

