% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\docType{data}
\name{crossoverCDFvector}
\alias{crossoverCDFvector}
\title{Contains recombination model information.}
\format{An object of class \code{logical} of length 1.}
\usage{
crossoverCDFvector
}
\description{
This vector contains the density between two recombination events, as a cumulative density function.
}
\keyword{datasets}

