% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{downloadGeneticMap}
\alias{downloadGeneticMap}
\title{Downloads a genetic map for a particular chromosome under GRCh37 coordinates for use with sim1000G.}
\usage{
downloadGeneticMap(chromosome, dir = NA)
}
\arguments{
\item{chromosome}{Chromosome number to download recombination distances from.}

\item{dir}{Directory to save the genetic map to (default: extdata)}
}
\description{
Downloads a genetic map for a particular chromosome under GRCh37 coordinates for use with sim1000G.
}
\examples{



downloadGeneticMap(22)


}

