% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{generateFakeGeneticMap}
\alias{generateFakeGeneticMap}
\title{Generates a fake genetic map.}
\usage{
generateFakeGeneticMap()
}
\description{
Generates a fake genetic map by approximating 1 cm / Mbp. Only used for examples.
}

