% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{generateRecombinationDistances_noInterference}
\alias{generateRecombinationDistances_noInterference}
\title{Generate recombination distances using a no-interference model.}
\usage{
generateRecombinationDistances_noInterference(n)
}
\arguments{
\item{n}{Number of distances to generate}
}
\value{
recombination distances in centimorgan
}
\description{
Generate recombination distances using a no-interference model.
}
\examples{

library("sim1000G")
mean ( generateRecombinationDistances_noInterference ( 200 ) )

}

