% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{generateSingleRecombinationVector}
\alias{generateSingleRecombinationVector}
\title{Genetates a recombination vector arising from one meiotic event.
The origin of segments is coded as (0 - haplotype1 ,  1 - haplotype2 )}
\usage{
generateSingleRecombinationVector(cm)
}
\arguments{
\item{cm}{The length of the region that we want to generate recombination distances.}
}
\description{
Genetates a recombination vector arising from one meiotic event.
The origin of segments is coded as (0 - haplotype1 ,  1 - haplotype2 )
}

