% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\docType{data}
\name{geneticMap}
\alias{geneticMap}
\title{Holds the genetic map information that is used for simulations.}
\format{An object of class \code{environment} of length 0.}
\usage{
geneticMap
}
\description{
Holds the genetic map information that is used for simulations.
}
\keyword{datasets}
