% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{getCMfromBP}
\alias{getCMfromBP}
\title{Converts centimorgan position to base-pair.}
\usage{
getCMfromBP(bp)
}
\arguments{
\item{bp}{vector of base-pair positions}
}
\description{
Converts centimorgan position to base-pair.
}

