% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{newNuclearFamily}
\alias{newNuclearFamily}
\title{Simulates genotypes for 1 family with 1 offspring}
\usage{
newNuclearFamily(family_id)
}
\arguments{
\item{family_id}{What will be the family_id (for example: 100)}
}
\value{
family structure object
}
\description{
Simulates genotypes for 1 family with 1 offspring
}
\examples{
# Previously you should have run the function startSimulation
\dontrun{
fam1 = newNuclearFamily(10)
}
}

