% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{data}
\name{pkg.opts}
\alias{pkg.opts}
\title{Holds general package options}
\format{An object of class \code{environment} of length 1.}
\usage{
pkg.opts
}
\description{
Holds general package options
}
\keyword{datasets}
