% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{plotRegionalGeneticMap}
\alias{plotRegionalGeneticMap}
\title{Generates a plot of the genetic map for a specified region.}
\usage{
plotRegionalGeneticMap(bp)
}
\arguments{
\item{bp}{Vector of base-pair positions to generate a plot for}
}
\description{
The plot shows the centimorgan vs base-pair positions.
In addition it showt the locations of the markers that have been read.
}

