% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{printMatrix}
\alias{printMatrix}
\title{Utility function that prints a matrix. Useful for IBD12 matrices.}
\usage{
printMatrix(m)
}
\arguments{
\item{m}{Matrix to be printed}
}
\description{
Utility function that prints a matrix. Useful for IBD12 matrices.
}
\examples{

printMatrix (  matrix(runif(16), nrow=4) )
}

