% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{readGeneticMapFromFile}
\alias{readGeneticMapFromFile}
\title{Reads a genetic map to be used for simulations.}
\usage{
readGeneticMapFromFile(filelocation)
}
\arguments{
\item{filelocation}{Filename containing the genetic map}
}
\description{
The file must be contain the following columns in order: chromosome, basepaire, rate(not used), centimorgan
}

