% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readvcf.R
\name{readVCF}
\alias{readVCF}
\title{Read a vcf file, with options to filter out low or high frequency markers.}
\usage{
readVCF(filename = "haplosims/1.vcf", thin = 1, maxNumberOfVariants = 400,
  min_maf = 0.02, max_maf = NA)
}
\arguments{
\item{filename}{Input VCF file}

\item{thin}{How much to thin markers}

\item{maxNumberOfVariants}{Maximum number of variants to keep from region}

\item{min_maf}{Minimum allele frequency of markers to keep}

\item{max_maf}{Maximum allele frequency of markers to keep}
}
\value{
none
}
\description{
Read a vcf file, with options to filter out low or high frequency markers.
}
\examples{
 \dontrun{
library(sim1000G)
 }
}

