% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{resetSimulation}
\alias{resetSimulation}
\title{Removes all individuals that have been simulated and resets the simulator.}
\usage{
resetSimulation()
}
\value{
nothing
}
\description{
Removes all individuals that have been simulated and resets the simulator.
}
\examples{

resetSimulation()

}
