% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{setRecombinationModel}
\alias{setRecombinationModel}
\title{Set recombination model to either poisson (no interference) or chi-square.}
\usage{
setRecombinationModel(model)
}
\arguments{
\item{model}{Either "poisson" or "chisq"}
}
\description{
Set recombination model to either poisson (no interference) or chi-square.
}
\examples{


generateUniformGeneticMap()

do_plots = 0

setRecombinationModel("chisq")
if(do_plots == 1)
 hist(generateRecombinationDistances(100000),n=200)

setRecombinationModel("poisson")
if(do_plots == 1)
 hist(generateRecombinationDistances(100000),n=200)

}
