% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{data}
\name{SIM}
\alias{SIM}
\title{Holds data necessary for a simulation.}
\format{An object of class \code{environment} of length 7.}
\usage{
SIM
}
\description{
Holds data necessary for a simulation.
}
\keyword{datasets}
