% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1000G-package.r
\docType{package}
\name{sim1000G-package}
\alias{sim1000G}
\alias{sim1000G-package}
\title{Simulations of rare/common variants using haplotype data from 1000 genomes}
\description{
(generic description)
}

