% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{startSimulation}
\alias{startSimulation}
\title{Start the simulation}
\usage{
startSimulation(vcf, totalNumberOfIndividuals = 250, randomdata = 0)
}
\arguments{
\item{vcf}{Input vcf file of a region (can be .gz). Must contain phased data.}

\item{totalNumberOfIndividuals}{Maximum Number of individuals that will ever be generated}

\item{randomdata}{If 1, disregards the genotypes in the vcf file and generates markers that are not in LD. Generally do not use.}
}
\description{
Start the simulation
}
\examples{
\dontrun{
library("sim1000G")
library(gplots)

vcf = readVCF("./region.vcf.gz", maxNumberOfVariants = 300 , min_maf = 0.12 ,max_maf = NA)

downloadGeneticMap( chromosome  = 4)
readGeneticMap( chromosome = 4)

plotRegionalGeneticMap(vcf$vcf[,2]+1)

startSimulation(vcf, totalNumberOfIndividuals = 1200)
}
}

