% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeped.R
\name{writePED}
\alias{writePED}
\title{Writes a plink compatible PED/MAP file from the simulated genotypes}
\usage{
writePED(vcf, fam, filename = "out")
}
\arguments{
\item{vcf}{vcf object used in simulation}

\item{fam}{Individuals / families to be written}

\item{filename}{Basename of output files (.ped/.map will be added automatically)}
}
\description{
Writes a plink compatible PED/MAP file from the simulated genotypes
}

