% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{computePairIBD1}
\alias{computePairIBD1}
\title{Computes pairwise IBD1 for a specific pair of individuals.
See function computePairIBD12 for description.}
\usage{
computePairIBD1(i, j)
}
\arguments{
\item{i}{Index of first individual}

\item{j}{Index of second individual}
}
\value{
Mean IBD1 as computed from shared haplotypes
}
\description{
Computes pairwise IBD1 for a specific pair of individuals.
See function computePairIBD12 for description.
}

