% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{generateChromosomeRecombinationPositions}
\alias{generateChromosomeRecombinationPositions}
\title{Genetates a recombination vector arising from one meiotic event.
The origin of segments is coded as (0 - haplotype1 ,  1 - haplotype2 )}
\usage{
generateChromosomeRecombinationPositions(chromosomeLength = 500)
}
\arguments{
\item{chromosomeLength}{The length of the region in cm.}
}
\description{
Genetates a recombination vector arising from one meiotic event.
The origin of segments is coded as (0 - haplotype1 ,  1 - haplotype2 )
}
\examples{

library("sim1000G")

readGeneticMap(4)
generateChromosomeRecombinationPositions(500)

}
