% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{generateRecombinationDistances}
\alias{generateRecombinationDistances}
\title{Generate recombination distances using a chi-square model.}
\usage{
generateRecombinationDistances(n)
}
\arguments{
\item{n}{Number of distances to generate}
}
\value{
vector of recombination distances in centimorgan
}
\description{
Generate recombination distances using a chi-square model.
}

