% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{loadSimulation}
\alias{loadSimulation}
\title{Load some previously saved simulation data by function saveSimulation}
\usage{
loadSimulation(id)
}
\arguments{
\item{id}{Name the simulation to load which was previously saved by saveSimulation}
}
\description{
Load some previously saved simulation data by function saveSimulation
}
\examples{

\dontrun{

loadSimulation("sim1")

}

}
