% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{readGeneticMap}
\alias{readGeneticMap}
\title{Reads a genetic map downloaded from the function downloadGeneticMap.}
\usage{
readGeneticMap(chromosome, dir = ".")
}
\arguments{
\item{chromosome}{Chromosome number to download recombination distances from.}

\item{dir}{Directory the map file is located.}
}
\description{
The map contains a complete chromosome to be used for simulations.
The file must be downloaded using the function downloadGeneticMap.
}

