% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{readGeneticMap}
\alias{readGeneticMap}
\title{Reads a genetic map downloaded from the function downloadGeneticMap.}
\usage{
readGeneticMap(chromosome, dir = NA)
}
\arguments{
\item{chromosome}{Chromosome number to download recombination distances from.}

\item{dir}{Directory the map file is located.}
}
\description{
The map must contains a complete chromosome or enough markers to cover the area that
will be simulated.
}
\details{
The file can be downloaded from the provided genetic maps by using the function downloadGeneticMap.
}
\examples{




readGeneticMap(chromosome = 22)



}

