% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{setRecombinationModel}
\alias{setRecombinationModel}
\title{Set options of the recombination model.}
\usage{
setRecombinationModel(model)
}
\arguments{
\item{model}{Either poisson or chisq}
}
\description{
Set options of the recombination model.
}
\examples{

setRecombinationModel("poisson")

setRecombinationModel("chisq")

}
