% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1000G-package.r
\docType{package}
\name{sim1000G-package}
\alias{sim1000G}
\alias{sim1000G-package}
\title{Simulations of rare/common variants using haplotype data from 1000 genomes}
\description{
Documentation and examples can be found at the package directory folder inst / doc
or at our github url:
https://adimitromanolakis.github.io/sim1000G/
inst/doc/SimulatingFamilyData.html
}
\details{
See also our github repository page at:
https://github.com/adimitromanolakis/sim1000G
}
