% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readvcf.R
\name{subsetVCF}
\alias{subsetVCF}
\title{Generate a market subset of a vcf file}
\usage{
subsetVCF(vcf, varid)
}
\arguments{
\item{vcf}{VCF data as created by function readVCF}

\item{varid}{varid of markers to subset. Should be a selection from vcf$varid}
}
\value{
VCF object to be used by startSimulation function.
}
\description{
Generate a market subset of a vcf file
}
\examples{
 \dontrun{
library(sim1000G)
 }
}

